[GAL-104] Split Gav and Ga from ArtifactCoords
