+++ cli/src/main/java/org/jboss/galleon/cli/AbstractStateCommand.java
@@ -20,6 +20,7 @@ import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Collections;
+
 import org.aesh.command.impl.completer.FileOptionCompleter;
 import org.aesh.command.impl.internal.ParsedCommand;
 import org.aesh.command.option.Option;
