+++ cli/src/main/java/org/jboss/galleon/cli/ChangesCommand.java
@@ -21,6 +21,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.Map;
+
 import org.aesh.command.CommandDefinition;
 import org.aesh.command.option.Option;
 import org.aesh.io.Resource;
