+++ cli/src/main/java/org/jboss/galleon/cli/PmSession.java
@@ -19,6 +19,7 @@ package org.jboss.galleon.cli;
 import java.io.PrintStream;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+
 import org.aesh.command.activator.CommandActivator;
 import org.aesh.command.activator.CommandActivatorProvider;
 import org.aesh.command.activator.OptionActivator;
