+++ cli/src/main/java/org/jboss/galleon/cli/ProvisionedSpecCommand.java
@@ -18,6 +18,7 @@ package org.jboss.galleon.cli;
 
 import java.util.ArrayList;
 import java.util.List;
+
 import org.aesh.command.Command;
 import org.aesh.command.CommandException;
 import org.aesh.command.CommandResult;
