+++ cli/src/main/java/org/jboss/galleon/cli/ProvisionedSpecExportCommand.java
@@ -19,6 +19,7 @@ package org.jboss.galleon.cli;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+
 import org.aesh.command.CommandDefinition;
 import org.aesh.command.option.Argument;
 import org.aesh.io.Resource;
