+++ core/src/main/java/org/jboss/galleon/config/FeaturePackDepsConfigBuilder.java
@@ -19,6 +19,7 @@ package org.jboss.galleon.config;
 import java.util.Collections;
 import java.util.Map;
 import java.util.Map.Entry;
+
 import org.jboss.galleon.Errors;
 import org.jboss.galleon.ProvisioningDescriptionException;
 import org.jboss.galleon.ProvisioningException;
