+++ core/src/main/java/org/jboss/galleon/diff/FileSystemDiff.java
@@ -29,11 +29,10 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
+import java.util.Set;
 
 import org.jboss.galleon.MessageWriter;
 import org.jboss.galleon.util.PathFilter;
-
-import java.util.Set;
 
 import difflib.DiffUtils;
 import difflib.Patch;
