+++ core/src/main/java/org/jboss/galleon/runtime/FeaturePackRuntimeBuilder.java
@@ -40,8 +40,8 @@ import org.jboss.galleon.type.ParameterTypeProvider;
 import org.jboss.galleon.type.builtin.BuiltInParameterTypeProvider;
 import org.jboss.galleon.universe.FeaturePackLocation.FPID;
 import org.jboss.galleon.universe.FeaturePackLocation.ProducerSpec;
-import org.jboss.galleon.util.LayoutUtils;
 import org.jboss.galleon.util.CollectionUtils;
+import org.jboss.galleon.util.LayoutUtils;
 import org.jboss.galleon.xml.FeatureGroupXmlParser;
 import org.jboss.galleon.xml.FeatureSpecXmlParser;
 import org.jboss.galleon.xml.PackageXmlParser;
