+++ core/src/main/java/org/jboss/galleon/runtime/ProvisioningRuntime.java
@@ -28,6 +28,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
+
 import javax.xml.stream.XMLStreamException;
 
 import org.jboss.galleon.ArtifactCoords;
