+++ core/src/main/java/org/jboss/galleon/runtime/ResolvedFeatureId.java
@@ -20,8 +20,8 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 
-import org.jboss.galleon.ArtifactCoords;
 import org.jboss.galleon.Constants;
+import org.jboss.galleon.Gav;
 import org.jboss.galleon.ProvisioningDescriptionException;
 import org.jboss.galleon.universe.FeaturePackLocation;
 import org.jboss.galleon.universe.FeaturePackLocation.ProducerSpec;
@@ -158,7 +158,7 @@ public class ResolvedFeatureId {
                 } else if(version == null) {
                     formatException(str);
                 } else {
-                    specId = new ResolvedSpecId(LegacyGalleon1Universe.toFpl(ArtifactCoords.newGav(groupId, artifactId, version)).getProducer(), buf.toString());
+                    specId = new ResolvedSpecId(LegacyGalleon1Universe.toFpl(new Gav(groupId, artifactId, version)).getProducer(), buf.toString());
                     break;
                 }
                 buf.setLength(0);
