+++ core/src/main/java/org/jboss/galleon/universe/galleon1/LegacyGalleon1FeaturePackInstaller.java
@@ -18,6 +18,7 @@
 package org.jboss.galleon.universe.galleon1;
 
 import java.nio.file.Path;
+
 import org.jboss.galleon.ArtifactRepositoryManager;
 import org.jboss.galleon.ProvisioningException;
 import org.jboss.galleon.universe.FeaturePackLocation;
