+++ core/src/main/java/org/jboss/galleon/xml/PackageXmlParser10.java
@@ -19,6 +19,7 @@ package org.jboss.galleon.xml;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
+
 import javax.xml.namespace.QName;
 import javax.xml.stream.XMLStreamConstants;
 import javax.xml.stream.XMLStreamException;
