+++ core/src/main/java/org/jboss/galleon/xml/ProvisioningXmlParser10.java
@@ -20,11 +20,12 @@ import java.util.EnumSet;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
+
 import javax.xml.namespace.QName;
 import javax.xml.stream.XMLStreamConstants;
 import javax.xml.stream.XMLStreamException;
 
-import org.jboss.galleon.ArtifactCoords;
+import org.jboss.galleon.Gav;
 import org.jboss.galleon.ProvisioningDescriptionException;
 import org.jboss.galleon.config.ConfigCustomizationsBuilder;
 import org.jboss.galleon.config.ConfigId;
@@ -246,7 +247,7 @@ public class ProvisioningXmlParser10 implements PlugableXmlParser<ProvisioningCo
             throw ParsingUtils.missingAttributes(reader.getLocation(), required);
         }
         String origin = null;
-        final FeaturePackConfig.Builder depBuilder = FeaturePackConfig.builder(LegacyGalleon1Universe.toFpl(ArtifactCoords.newGav(groupId, artifactId, version)));
+        final FeaturePackConfig.Builder depBuilder = FeaturePackConfig.builder(LegacyGalleon1Universe.toFpl(new Gav(groupId, artifactId, version)));
         while (reader.hasNext()) {
             switch (reader.nextTag()) {
                 case XMLStreamConstants.END_ELEMENT: {
